+++ src/ImmichFolderWatch.Gui/Services/ConfigVerificationRunner.cs
@@ -1,4 +1,3 @@
-using System.Reflection;
 using ImmichFolderWatch.Core.Configuration;
 using ImmichFolderWatch.Core.Models;
 using ImmichFolderWatch.Immich;
@@ -54,24 +53,6 @@ internal sealed class ConfigVerificationRunner
             : VerificationResult.Passed();
     }
 
-    private static string GetProductVersion()
-    {
-        var informationalVersion = typeof(ConfigVerificationRunner)
-            .Assembly
-            .GetCustomAttribute<AssemblyInformationalVersionAttribute>()
-            ?.InformationalVersion;
-
-        if (!string.IsNullOrWhiteSpace(informationalVersion))
-        {
-            var plusIndex = informationalVersion.IndexOf('+');
-            return plusIndex >= 0
-                ? informationalVersion[..plusIndex]
-                : informationalVersion;
-        }
-
-        return "1.2.1";
-    }
-
     private static HttpClient CreateHttpClient(AppConfig normalizedConfig)
     {
         var httpClient = new HttpClient
@@ -81,7 +62,7 @@ internal sealed class ConfigVerificationRunner
         };
 
         httpClient.DefaultRequestHeaders.Add("x-api-key", normalizedConfig.Immich.ApiKey);
-        httpClient.DefaultRequestHeaders.UserAgent.ParseAdd($"immich-folder-watch-gui/{GetProductVersion()}");
+        httpClient.DefaultRequestHeaders.UserAgent.ParseAdd($"immich-folder-watch-gui/{ProductVersionProvider.GetProductVersion()}");
         return httpClient;
     }